\name{fit.models.registry}
\alias{fit.models.registry}
\docType{data}

\title{fit.models Registry}

\description{
A list containing information on comparable models.
}

\details{
R's package system requires a help file for this object but ideally it should be invisible. Do not attempt to make changes to this list directly, rather use hte functions \code{\link{fm.register.vclass}} and \code{\link{fm.register.class.in.vclass}}.
}

\keyword{misc}
