% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmclass.register.R
\name{fmclass.register}
\alias{fmclass.register}
\alias{fmclass.add.class}
\alias{fmclass.add.class}
\title{Register Comparable Functions}
\usage{
fmclass.register(fmclass, classes, validation.function = NULL)

fmclass.add.class(fmclass, class, warn = TRUE)
}
\arguments{
\item{fmclass}{a character string naming the fit.models class to be added.}

\item{classes}{a character vector naming one or more classes that can be
compared by the methods defined for the fit.models class in \code{fmclass}.}

\item{validation.function}{a function returning \code{TRUE} when the models
are comparable.}

\item{class}{a character string specifying a class compatible with the
methods of \code{fmclass}.}

\item{warn}{a logical value. If TRUE, a warning is printed if \code{class}
is already registered.}
}
\value{
a null value is invisibly returned.
}
\description{
The fit.models package maintains a list of comparable models.
These functions provide an api to modify this list.
}
\details{
See the package vignette.
}
\keyword{misc}
