% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexPlot.lmfm.R
\name{indexPlot.lmfm}
\alias{indexPlot.lmfm}
\title{Comparison Index (Time) Plot}
\usage{
indexPlot.lmfm(x, fun, level = 0.95, id.n = 3, ...)
}
\arguments{
\item{x}{an lmfm object.}

\item{fun}{a function to extract the appropriate quantity from \code{x}.}

\item{level}{a numeric value between 0 and 1 specifying the confidence level
used to draw the threshold in the plot.}

\item{id.n}{a non-negative integer value specifying the number of extreme
points to identify.}

\item{\dots}{any additional arguments are passed to
\code{\link[lattice]{xyplot}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces side-by-side index (time) plots.
}
\keyword{hplot}
