% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaidKernDenPlot.lmfm.R
\name{overlaidKernDenPlot.lmfm}
\alias{overlaidKernDenPlot.lmfm}
\title{Overlaid Kernel Density Estimate Plot}
\usage{
overlaidKernDenPlot.lmfm(x, fun, ...)
}
\arguments{
\item{x}{an lmfm object.}

\item{fun}{a function to extract the appropriate quantity from \code{x}.}

\item{\dots}{additional arguments are passed to
\code{\link[lattice]{densityplot}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces an overlaid kernel density plot.
}
\keyword{hplot}
