% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaidKernelDensityPlot.R
\name{overlaidKernelDensityPlot}
\alias{overlaidKernelDensityPlot}
\title{Overlaid Kernel Density Estimate Plot}
\usage{
overlaidKernelDensityPlot(x, fun, ...)
}
\arguments{
\item{x}{a \code{fit.models} object.}

\item{fun}{a function to extract the desired quantity from \code{x}.}

\item{\dots}{additional arguments are passed to
\code{densityplot}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces an overlaid kernel density plot.
}
\keyword{hplot}
