% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaidQQPlot.lmfm.R
\name{overlaidQQPlot.lmfm}
\alias{overlaidQQPlot.lmfm}
\title{Overlaid Normal QQ Plot}
\usage{
overlaidQQPlot.lmfm(x, fun, ...)
}
\arguments{
\item{x}{an lmfm object.}

\item{fun}{a function to extract the appropriate quantity from \code{x}.}

\item{\dots}{additional arguments are passed to
\code{\link[lattice]{qqmath}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces an overlaid normal QQ plot.
}
\keyword{hplot}
