% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaidQQPlot.R
\name{overlaidQQPlot}
\alias{overlaidQQPlot}
\title{Overlaid Normal QQ Plot}
\usage{
overlaidQQPlot(x, fun, ...)
}
\arguments{
\item{x}{a \code{fit.models} object.}

\item{fun}{a function to extract the desired quantity from \code{x}.}

\item{\dots}{additional arguments are passed to
\code{\link[lattice]{qqmath}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces an overlaid normal QQ plot.
}
\keyword{hplot}
