% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaidSimpleRegressionPlot.R
\name{overlaidSimpleRegressionPlot}
\alias{overlaidSimpleRegressionPlot}
\title{Scatter Plot with Overlaid Fits}
\usage{
overlaidSimpleRegressionPlot(x, lwd.reg, col.reg, ...)
}
\arguments{
\item{x}{a \code{fit.models} object.}

\item{lwd.reg}{a vector with length equal to the number of models in
\code{x} specifying the line widths used in the plot.}

\item{col.reg}{a vector with length equal to the number of models in
\code{x} specifying the line colors used in the plot.}

\item{\dots}{additional arguments are passed to
\code{\link[lattice]{xyplot}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces a scatter plot of the data with overlaid fits.
}
\keyword{hplot}
