% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.covfm.R
\name{plot.covfm}
\alias{plot.covfm}
\title{Plot Method}
\usage{
\method{plot}{covfm}(x, which.plots = 1:4, ...)
}
\arguments{
\item{x}{a \code{covfm} object.}

\item{which.plots}{either \code{"ask"} (character string) or an integer vector specifying
which plots to draw.  The plot options are (1) Mahalanobis Distance,
(2) Ellipses Matrix, (3) Screeplot (Eigenvalues of Covariance Estimate),
and (4) Distance - Distance Plot.}

\item{...}{additional arguments are passed to the plot subfunctions.}
}
\value{
\code{x} is returned invisibly.
}
\description{
Generic plot method for \dQuote{covfm} objects.
}
