% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.lmfm.R
\name{scatterPlot.lmfm}
\alias{scatterPlot.lmfm}
\title{Comparison Scatter Plot}
\usage{
scatterPlot.lmfm(object, x.fun, y.fun, ...)
}
\arguments{
\item{object}{an lmfm object.}

\item{x.fun}{a function to extract the x-axis quantity.}

\item{y.fun}{a function to extract the y-axis quantity.}

\item{\dots}{additional arguments.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces a side-by-side scatter plot.
}
\keyword{hplot}
