% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screePlot.covfm.R
\name{screePlot.covfm}
\alias{screePlot.covfm}
\title{Comparison Screeplot}
\usage{
screePlot.covfm(x, npcs, strip = "", ...)
}
\arguments{
\item{x}{a \code{"covfm"} object.}

\item{npcs}{a postive integer value specifying the number of components to be plotted.}

\item{strip}{a character string printed in the \dQuote{strip} at the top of the plot.}

\item{...}{additional arguments are passed to \code{xyplot}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Overlaid screeplot for the models in a \dQuote{covfm} object.
}
