% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideBySideKernelDensityPlot.R
\name{sideBySideKernelDensityPlot}
\alias{sideBySideKernelDensityPlot}
\title{Comparison Kernel Density Estimate Plot}
\usage{
sideBySideKernelDensityPlot(x, fun, ...)
}
\arguments{
\item{x}{a \code{fit.models} object.}

\item{fun}{a function to extract the desired quantity from \code{x}.}

\item{\dots}{additional arguments are passed to
\code{\link[lattice]{xyplot}}.}
}
\value{
the \code{trellis} object is invisibly returned.
}
\description{
Produces side-by-side kernel density estimate plots.
}
\keyword{hplot}
