% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.covfm.R
\name{summary.covfm}
\alias{summary.covfm}
\title{Summary Method}
\usage{
\method{summary}{covfm}(object, corr = FALSE, ...)
}
\arguments{
\item{object}{a \dQuote{covfm} object.}

\item{corr}{a logical value passed as an attribute to the \code{print}
method. When \code{TRUE}, correlations are compared in the
textual output.}

\item{\dots}{additional arguments for the summary method.}
}
\description{
Generic summary method for \dQuote{covfm} objects.
}
