% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfm.R
\name{vcor}
\alias{vcor}
\title{Calculate Correlation Matrix for a Fitted Model Object}
\usage{
vcor(object, ...)
}
\arguments{
\item{object}{a fitted model object, typically.  Sometimes
also a \code{summary()} object of such a fitted
model.}

\item{\dots}{additional arguments for method functions.}
}
\description{
Retrieve a correlation matrix estimate from a
             fitted model object. The default method uses
             \code{\link[stats]{cov2cor}} to scale the
             covariance matrix returned by
             \code{\link[stats]{vcov}}.
}
