% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfm.R
\name{center}
\alias{center}
\title{Calculate Location Estimate for a Fitted Model Object}
\usage{
center(object, ...)
}
\arguments{
\item{object}{a fitted model object, typically.  Sometimes
also a \code{summary()} object of such a fitted
model.}

\item{\dots}{additional arguments for method functions.}
}
\description{
Returns the location estimated from a
             location/scatter-type fitted model object.
}
