\name{overlaidKernDenPlot.lmfm}
\alias{overlaidKernDenPlot.lmfm}

\title{Overlaid Kernel Density Estimate Plot}

\description{
Produces an overlaid kernel density plot.
}

\usage{
overlaidKernDenPlot.lmfm(x, fun, \dots)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{fun}{a function to extract the appropriate quantity from \code{x}.}

  \item{\dots}{additional arguments are passed to \code{\link[lattice]{densityplot}}.}
}

\value{
the \code{trellis} object is invisibly returned.
}

\keyword{hplot}


