\name{scatterPlot.lmfm}
\alias{scatterPlot.lmfm}

\title{Comparison Scatter Plot}

\description{
Produces a side-by-side scatter plot.
}

\usage{
scatterPlot.lmfm(object, x.fun, y.fun, \dots)
}

\arguments{
  \item{object}{an lmfm object.}
  \item{x.fun}{a function to extract the x-axis quantity.}
  \item{y.fun}{a function to extract the y-axis quantity.}
  \item{\dots}{additional arguments.}
}

\value{
  the \code{trellis} object is invisibly returned.
}

\keyword{hplot}


