% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.covfm.R
\name{summary.covfm}
\alias{summary.covfm}
\title{Summary Method}
\usage{
\method{summary}{covfm}(object, ...)
}
\arguments{
\item{object}{a \dQuote{covfm} object.}

\item{...}{additional arguments for the summary method.}
}
\description{
Generic summary method for \dQuote{covfm} objects.
}
